arm64: dts: msm: Increase GLINK intents
There is currently a potential for a deadlock in the DSP when it wants to send a GLINK message but is waiting for an intent.

Increase the number of intents to 1 more than the number of DSP supported threads so that the DSP won't have to wait for intents.

Signed-off-by: Liam Mark <quic_lmark@quicinc.com>
Change-Id: I8503886606027df45d6bcd2f4492a484cdc84696
(cherry picked from commit a229c8ed3f0d1b66cf440d2509887ae2149859b7)